Create custommutefunction
This is a little bash script that I TAKE NO CREDIT FOR. I'm so sorry that i cannot find the random stackoverflow user who wrote this solution. If you want to claim it or take it down just contact me I have only kept it up because this is an extremely useful fix for a glitch within alsamixer that mutes all channels on mute but only unmutes the master channel when unmuted.