Create init.vim
Again a lot of this file was taken from other sources, mainly this article: https://jdhao.github.io/2018/12/24/centos_nvim_install_use_guide_en/ .If this breaks copyright rules and you want to take this down just message me. This is the config file for neovim, my preference of text editor for programming.