Create colors
Generated by pywal. Ideally I aim to unify all config files to get a colourscheme from one text file, generated by wal from one image, making it easy to switch between themes. This would mean I could tune the colours to match the wallpaper, as although pywal is incredible it doesnt get it right 100% of the time. Basically I wanted this so i could have more control, and to make it easier to expand in the future with eg an auto light and dark scheme based on time.